MultiMotionFusion
MultiMotionFusion implements an online tracking and modelling approach for
multiple rigid objects, including the environment. It enables the
reconstruction and pose estimation of previously unseen objects with
respect to the world and the camera. The segmentation of the scene and the
detection of new objects relies on motion and thus does not require prior
information about objects or the environment.

https://doi.org/10.48550/arXiv.2204.11923